Move ADC Filter and Center Beep settings to 'Other' page.
